// src/types.ts
export type TransactionType = "entrada" | "saida";